feat: 입력 시 참고할 수 있는 문구 추가
