+++ two_week/daeun/AdvancedTask/AdvancedTask/RequiredTask/InputError.swift
@@ -1,11 +1,9 @@
 //
-//  ConversionError.swift
+//  InputError.swift
 //  AdvancedTask
 //
 //  Created by 곽다은 on 3/18/25.
 //
-
-import Foundation
 
 enum InputError: Error {
     case emptyInput
