+++ two_week/daeun/AdvancedTask/AdvancedTask/RequiredTask/RequiredTask1.swift
@@ -15,4 +15,5 @@ func calculate(_ operate: (Int, Int) -> String) {
 func testCalculate() {
     print("-- calculate(_:) --")
     calculate(sum)
+    waitForEnter()
 }
