+++ two_week/daeun/AdvancedTask/AdvancedTask/RequiredTask/RequiredTask2.swift
@@ -9,7 +9,7 @@
 func testForInForEach() {
     print("-- for-in ForEach --")
     
-    let numbers: [Int] = getValidArrayInput()
+    let numbers: [Int] = getValidArrayInput(reference: "for-in과 forEach의 결과값이 같은지 검사합니다.")
     var forInResult: [String] = []
     var forEachResult: [String] = []
     
@@ -23,17 +23,19 @@ func testForInForEach() {
     
     print("for-in 결과: \(forInResult)")
     print("forEach 결과: \(forEachResult)")
+    waitForEnter()
 }
 
 // MARK: higher order function
 func testHighOrderFunctionChaining() {
     print("-- Higher Order Function Chaining --")
     
-    let input: [Int] = getValidArrayInput()
+    let input: [Int] = getValidArrayInput(reference: "고차함수를 통해 짝수인 요소가 필터링됩니다.")
     let output = input.filter { $0 % 2 == 0 }.map { String($0) }
     
     print("input: \(input)")
     print("output: \(output)")
+    waitForEnter()
 }
 
 // MARK: custom higher order function
@@ -48,8 +50,9 @@ func testCreateCustomHigherOrderFunction() {
         return convertedArray
     }
 
-    let input: [Int] = getValidArrayInput()
+    let input: [Int] = getValidArrayInput(reference: "정수 배열을 문자열 배열로 변환합니다.")
     let myMapResult = myMap(input) { String($0) }
     
     print("myMap result: \(myMapResult)")
+    waitForEnter()
 }
